(PC-7477) Create SeeItineraryButton
